+++ babel.config.js
@@ -21,7 +21,21 @@ module.exports = {
     ],
     [
       "@babel/plugin-proposal-class-properties"
-    ]
+    ],
+	["module-resolver", {
+	  "root": ["./src"],
+	  "alias": {
+		"@uixkit.react/components": "./src/client/components",
+		"@uixkit.react/plugins": "./src/client/components/_third-party-plugins",
+		"@uixkit.react/router": "./src/client/router",
+		"@uixkit.react/reducers": "./src/client/reducers",
+		"@uixkit.react/pages": "./src/client/views/_pages",
+		"@uixkit.react/actions": "./src/client/actions",
+		"@uixkit.react/server": "./src/server",
+		"@uixkit.react/store": "./src/store"
+	  }
+	}]
     
   ]
 };
+