+++ src/client/components/Buttons/index.js
@@ -9,25 +9,25 @@ import React, { Component, Fragment } from 'react';
 
 /*-- Apply Third-party plugins --*/
 //1) Bootstrap
-import '../_third-party-plugins/Bootstrap/scss/bootstrap.scss';  
+import '@uixkit.react/plugins/Bootstrap/scss/bootstrap.scss';  
 
 
 //2) Fontawesome
-import '../_third-party-plugins/FontAwesome/scss/fontawesome.scss';  
-import '../_third-party-plugins/FontAwesome/scss/v4-shims.scss';  
-import '../_third-party-plugins/FontAwesome/scss/solid.scss';  
-import '../_third-party-plugins/FontAwesome/scss/brands.scss';  
-import '../_third-party-plugins/FontAwesome/scss/regular.scss';  
+import '@uixkit.react/plugins/FontAwesome/scss/fontawesome.scss';  
+import '@uixkit.react/plugins/FontAwesome/scss/v4-shims.scss';  
+import '@uixkit.react/plugins/FontAwesome/scss/solid.scss';  
+import '@uixkit.react/plugins/FontAwesome/scss/brands.scss';  
+import '@uixkit.react/plugins/FontAwesome/scss/regular.scss';  
 
 
 
 /*-- Apply global scripts and styles --*/
-import '../_utilities/scss/_all.scss';
-import '../_utilities/scss/rtl/_all-rtl.scss';
-import { __ } from '../_utilities/js/_all.js';
+import '@uixkit.react/components/_utilities/scss/_all.scss';
+import '@uixkit.react/components/_utilities/scss/rtl/_all-rtl.scss';
+import { __ } from '@uixkit.react/components/_utilities/js/_all.js';
 
 /*-- Apply this component styles --*/
-import './styles/index.scss';
+import '@uixkit.react/components/Buttons/styles/index.scss';
 
 
 const styles = {