+++ src/client/components/_utilities/js/_all.js
@@ -7,7 +7,7 @@
 
 
 */
-import __ from "./shortcut.js";
+import __ from '@uixkit.react/components/_utilities/js/shortcut.js';
 
 
 export {