+++ src/client/views/_pages/NestedRoutes/index.js
@@ -5,12 +5,11 @@ import {
 	NavLink,
 	useRouteMatch,
 	useParams
-} from "react-router-dom";
-import customRoutesConfig from '../../../router/RoutesConfig.js';
-import Header from '../../../components/Header/index.js';
-import Footer from '../../../components/Footer/index.js';
+} from 'react-router-dom';
+import customRoutesConfig from '@uixkit.react/router/RoutesConfig.js';
+import Footer from '@uixkit.react/components/Footer/index.js';
 
-import NestedRoutesDetail from "./NestedRoutesDetail.js";
+import NestedRoutesDetail from '@uixkit.react/pages/NestedRoutes/NestedRoutesDetail.js';
 
 
 export default (props) => {