+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Changelog
 
+#### = 0.0.18 (April 18, 2021) =
+
+* Supports directories and custom aliases for specific files when the express server is running.
+* Detail optimization.
+
 
 #### = 0.0.17 (April 15, 2021) =
 
 * Add Nested Routes demo.
 * Add more UI components.
 * Minor adjustment of the Router directory structure. 
+
 
 #### = 0.0.16 (April 13, 2021) =